requirements.txt
peewee~=3.14.4
schedule~=1.1.0
PyMySQL~=1.0.2
selenium~=3.141.0
beautifulsoup4~=4.9.3
lxml~=4.6.3
fake-useragent~=0.1.11
Django==3.2.8
djangorestframework==3.12.4
psycopg2==2.9.2
celery==5.1.2
apscheduler==3.7.0
requests~=2.26.0